Refactor valueFromNode
